+++ src/pandas_indexing/__init__.py
@@ -15,6 +15,7 @@ from .core import (
     index_names,
     projectlevel,
     semijoin,
+    uniquelevel,
 )
 from .selectors import isin, ismatch